+++ opm/autodiff/BlackoilModelBase_impl.hpp
@@ -25,6 +25,7 @@
 #define OPM_BLACKOILMODELBASE_IMPL_HEADER_INCLUDED
 
 #include <opm/autodiff/BlackoilModelBase.hpp>
+#include <opm/autodiff/BlackoilDetails.hpp>
 
 #include <opm/autodiff/AutoDiffBlock.hpp>
 #include <opm/autodiff/AutoDiffHelpers.hpp>
@@ -92,73 +93,6 @@ typedef Eigen::Array<double,
                      Eigen::Dynamic,
                      Eigen::Dynamic,
                      Eigen::RowMajor> DataBlock;
-
-
-namespace detail {
-
-
-    inline
-    std::vector<int>
-    buildAllCells(const int nc)
-    {
-        std::vector<int> all_cells(nc);
-
-        for (int c = 0; c < nc; ++c) { all_cells[c] = c; }
-
-        return all_cells;
-    }
-
-
-
-    template <class PU>
-    std::vector<bool>
-    activePhases(const PU& pu)
-    {
-        const int maxnp = Opm::BlackoilPhases::MaxNumPhases;
-        std::vector<bool> active(maxnp, false);
-
-        for (int p = 0; p < pu.MaxNumPhases; ++p) {
-            active[ p ] = pu.phase_used[ p ] != 0;
-        }
-
-        return active;
-    }
-
-
-
-    template <class PU>
-    std::vector<int>
-    active2Canonical(const PU& pu)
-    {
-        const int maxnp = Opm::BlackoilPhases::MaxNumPhases;
-        std::vector<int> act2can(maxnp, -1);
-
-        for (int phase = 0; phase < maxnp; ++phase) {
-            if (pu.phase_used[ phase ]) {
-                act2can[ pu.phase_pos[ phase ] ] = phase;
-            }
-        }
-
-        return act2can;
-    }
-
-
-
-    inline
-    double getGravity(const double* g, const int dim) {
-        double grav = 0.0;
-        if (g) {
-            // Guard against gravity in anything but last dimension.
-            for (int dd = 0; dd < dim - 1; ++dd) {
-                assert(g[dd] == 0.0);
-            }
-            grav = g[dim - 1];
-        }
-        return grav;
-    }
-
-} // namespace detail
-
 
     template <class Grid, class WellModel, class Implementation>
     BlackoilModelBase<Grid, WellModel, Implementation>::
@@ -1087,135 +1021,6 @@ namespace detail {
     {
         return linsolver_.computeNewtonIncrement(residual_);
     }
-
-
-
-
-
-    namespace detail
-    {
-        /// \brief Compute the L-infinity norm of a vector
-        /// \warn This function is not suitable to compute on the well equations.
-        /// \param a The container to compute the infinity norm on.
-        ///          It has to have one entry for each cell.
-        /// \param info In a parallel this holds the information about the data distribution.
-        inline
-        double infinityNorm( const ADB& a, const boost::any& pinfo = boost::any() )
-        {
-            static_cast<void>(pinfo); // Suppress warning in non-MPI case.
-#if HAVE_MPI
-            if ( pinfo.type() == typeid(ParallelISTLInformation) )
-            {
-                const ParallelISTLInformation& real_info =
-                    boost::any_cast<const ParallelISTLInformation&>(pinfo);
-                double result=0;
-                real_info.computeReduction(a.value(), Reduction::makeLInfinityNormFunctor<double>(), result);
-                return result;
-            }
-            else
-#endif
-            {
-                if( a.value().size() > 0 ) {
-                    return a.value().matrix().lpNorm<Eigen::Infinity> ();
-                }
-                else { // this situation can occur when no wells are present
-                    return 0.0;
-                }
-            }
-        }
-
-        /// \brief Compute the Euclidian norm of a vector
-        /// \warning In the case that num_components is greater than 1
-        ///          an interleaved ordering is assumed. E.g. for each cell
-        ///          all phases of that cell are stored consecutively. First
-        ///          the ones for cell 0, then the ones for cell 1, ... .
-        /// \param it              begin iterator for the given vector
-        /// \param end             end iterator for the given vector
-        /// \param num_components  number of components (i.e. phases) in the vector
-        /// \param pinfo           In a parallel this holds the information about the data distribution.
-        template <class Iterator>
-        inline
-        double euclidianNormSquared( Iterator it, const Iterator end, int num_components, const boost::any& pinfo = boost::any() )
-        {
-            static_cast<void>(num_components); // Suppress warning in the serial case.
-            static_cast<void>(pinfo); // Suppress warning in non-MPI case.
-#if HAVE_MPI
-            if ( pinfo.type() == typeid(ParallelISTLInformation) )
-            {
-                const ParallelISTLInformation& info =
-                    boost::any_cast<const ParallelISTLInformation&>(pinfo);
-                typedef typename Iterator::value_type Scalar;
-                Scalar product = 0.0;
-                int size_per_component = (end - it);
-                size_per_component /= num_components; // two lines to supresse unused warning.
-                assert((end - it) == num_components * size_per_component);
-
-                if( num_components == 1 )
-                {
-                    auto component_container =
-                        boost::make_iterator_range(it, end);
-                    info.computeReduction(component_container,
-                                           Opm::Reduction::makeInnerProductFunctor<double>(),
-                                           product);
-                }
-                else
-                {
-                    auto& maskContainer = info.getOwnerMask();
-                    auto mask = maskContainer.begin();
-                    assert(static_cast<int>(maskContainer.size()) == size_per_component);
-
-                    for(int cell = 0; cell < size_per_component; ++cell, ++mask)
-                    {
-                        Scalar cell_product = (*it) * (*it);
-                        ++it;
-                        for(int component=1; component < num_components;
-                            ++component, ++it)
-                        {
-                            cell_product += (*it) * (*it);
-                        }
-                        product += cell_product * (*mask);
-                    }
-                }
-                return info.communicator().sum(product);
-            }
-            else
-#endif
-            {
-                double product = 0.0 ;
-                for( ; it != end; ++it ) {
-                    product += ( *it * *it );
-                }
-                return product;
-            }
-        }
-
-        /// \brief Compute the L-infinity norm of a vector representing a well equation.
-        /// \param a The container to compute the infinity norm on.
-        /// \param info In a parallel this holds the information about the data distribution.
-        inline
-        double infinityNormWell( const ADB& a, const boost::any& pinfo )
-        {
-            static_cast<void>(pinfo); // Suppress warning in non-MPI case.
-            double result=0;
-            if( a.value().size() > 0 ) {
-                result = a.value().matrix().lpNorm<Eigen::Infinity> ();
-            }
-#if HAVE_MPI
-            if ( pinfo.type() == typeid(ParallelISTLInformation) )
-            {
-                const ParallelISTLInformation& real_info =
-                    boost::any_cast<const ParallelISTLInformation&>(pinfo);
-                result = real_info.communicator().max(result);
-            }
-#endif
-            return result;
-        }
-
-    } // namespace detail
-
-
-
-
 
     template <class Grid, class WellModel, class Implementation>
     void
